# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nulls Report production portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for a private attachment upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UploadUrlRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 255 }
        size: { type: number, minimum: 1, maximum: 52428800 }
        contentType: { type: string, minLength: 1, maxLength: 255 }
      required: [name, size, contentType]
    UploadUrlResponse:
      type: object
      properties:
        id: { type: number }
        uploadURL: { type: string }
        objectPath: { type: string }
      required: [id, uploadURL, objectPath]
...